updated classes to use Int instead of Long for api

+++ src/main/kotlin/com/seekheart/superpal/models/web/BossRequest.kt
@@ -1,3 +1,3 @@
 package com.seekheart.superpal.models.web
 
-data class BossRequest(val damage: Long)
+data class BossRequest(val damage: Int)
