+++ src/main/kotlin/com/seekheart/superpal/models/web/RaidRequest.kt
@@ -1,3 +1,3 @@
 package com.seekheart.superpal.models.web
 
-data class RaidRequest(val tier: Long, val state: RaidStatusOptions)
+data class RaidRequest(val tier: Int, val state: RaidStatusOptions)
